docs: 폴더 위치 변경 playlist에서 category로 이동
