+++ src/pages/categoryPage/CategoryPage.tsx
@@ -1,7 +1,73 @@
+import PlaylistCardList from '../../components/@common/card/PlaylistCardList';
+
 interface CategoryPageProps {}
 
+const predata = [
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Sunny Beach',
+    description: 'A beautiful sunny beach with crystal clear water.',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Mountain Landscape',
+    description: 'A breathtaking view of the mountain range during sunset.',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'City Skyline',
+    description: 'A panoramic view of a bustling city skyline at night.',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+    description: 'A serene pathway through a dense forest with tall trees.',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+    description: 'A serene pathway through a dense forest with tall trees.',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+    description: 'A serene pathway through a dense forest with tall trees.',
+  },
+];
+
+const predata2 = [
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Sunny Beach',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Mountain Landscape',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'City Skyline',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+  },
+  {
+    imgUrl: 'https://pbs.twimg.com/media/FYznstoXEAATx8k.jpg',
+    title: 'Forest Pathway',
+  },
+];
 const CategoryPage = ({}: CategoryPageProps) => {
-  return ;
+  return (
+    <>
+      <PlaylistCardList datas={predata} />
+      <PlaylistCardList datas={predata2} />
+    </>
+  );
 };
 
 export default CategoryPage;
